+++ tests/integration.rs
@@ -98,45 +98,45 @@ fn test_boolean_examples() {
 }
 
 #[test]
-fn test_with_configuration() {
-    let mut configuration = HashMapContext::new();
-    configuration.set_value("tr", Value::Boolean(true)).unwrap();
-    configuration.set_value("fa", Value::Boolean(false)).unwrap();
-    configuration.set_value("five", Value::Int(5)).unwrap();
-    configuration.set_value("six", Value::Int(6)).unwrap();
-    configuration.set_value("half", Value::Float(0.5)).unwrap();
-    configuration.set_value("zero", Value::Int(0)).unwrap();
+fn test_with_context() {
+    let mut context = HashMapContext::new();
+    context.set_value("tr", Value::Boolean(true)).unwrap();
+    context.set_value("fa", Value::Boolean(false)).unwrap();
+    context.set_value("five", Value::Int(5)).unwrap();
+    context.set_value("six", Value::Int(6)).unwrap();
+    context.set_value("half", Value::Float(0.5)).unwrap();
+    context.set_value("zero", Value::Int(0)).unwrap();
 
     assert_eq!(
-        eval_with_configuration("tr", &configuration),
+        eval_with_context("tr", &context),
         Ok(Value::Boolean(true))
     );
     assert_eq!(
-        eval_with_configuration("fa", &configuration),
+        eval_with_context("fa", &context),
         Ok(Value::Boolean(false))
     );
     assert_eq!(
-        eval_with_configuration("tr && false", &configuration),
+        eval_with_context("tr && false", &context),
         Ok(Value::Boolean(false))
     );
     assert_eq!(
-        eval_with_configuration("five + six", &configuration),
+        eval_with_context("five + six", &context),
         Ok(Value::Int(11))
     );
     assert_eq!(
-        eval_with_configuration("five * half", &configuration),
+        eval_with_context("five * half", &context),
         Ok(Value::Float(2.5))
     );
     assert_eq!(
-        eval_with_configuration("five < six && true", &configuration),
+        eval_with_context("five < six && true", &context),
         Ok(Value::Boolean(true))
     );
 }
 
 #[test]
 fn test_functions() {
-    let mut configuration = HashMapContext::new();
-    configuration.set_function(
+    let mut context = HashMapContext::new();
+    context.set_function(
         "sub2".to_string(),
         Function::new(
             Some(1),
@@ -151,34 +151,34 @@ fn test_functions() {
             }),
         ),
     ).unwrap();
-    configuration.set_value("five".to_string(), Value::Int(5)).unwrap();
+    context.set_value("five".to_string(), Value::Int(5)).unwrap();
 
     assert_eq!(
-        eval_with_configuration("sub2 5", &configuration),
+        eval_with_context("sub2 5", &context),
         Ok(Value::Int(3))
     );
     assert_eq!(
-        eval_with_configuration("sub2(5)", &configuration),
+        eval_with_context("sub2(5)", &context),
         Ok(Value::Int(3))
     );
     assert_eq!(
-        eval_with_configuration("sub2 five", &configuration),
+        eval_with_context("sub2 five", &context),
         Ok(Value::Int(3))
     );
     assert_eq!(
-        eval_with_configuration("sub2(five)", &configuration),
+        eval_with_context("sub2(five)", &context),
         Ok(Value::Int(3))
     );
     assert_eq!(
-        eval_with_configuration("sub2(3) + five", &configuration),
+        eval_with_context("sub2(3) + five", &context),
         Ok(Value::Int(6))
     );
 }
 
 #[test]
 fn test_n_ary_functions() {
-    let mut configuration = HashMapContext::new();
-    configuration.set_function(
+    let mut context = HashMapContext::new();
+    context.set_function(
         "sub2",
         Function::new(
             Some(1),
@@ -193,7 +193,7 @@ fn test_n_ary_functions() {
             }),
         ),
     ).unwrap();
-    configuration.set_function(
+    context.set_function(
         "avg",
         Function::new(
             Some(2),
@@ -211,7 +211,7 @@ fn test_n_ary_functions() {
             }),
         ),
     ).unwrap();
-    configuration.set_function(
+    context.set_function(
         "muladd",
         Function::new(
             Some(3),
@@ -233,57 +233,57 @@ fn test_n_ary_functions() {
             }),
         ),
     ).unwrap();
-    configuration.set_function(
+    context.set_function(
         "count",
         Function::new(
             None,
             Box::new(|arguments| Ok(Value::Int(arguments.len() as IntType))),
         ),
     ).unwrap();
-    configuration.set_value("five".to_string(), Value::Int(5)).unwrap();
+    context.set_value("five".to_string(), Value::Int(5)).unwrap();
 
     assert_eq!(
-        eval_with_configuration("avg(7, 5)", &configuration),
+        eval_with_context("avg(7, 5)", &context),
         Ok(Value::Int(6))
     );
     assert_eq!(
-        eval_with_configuration("avg(sub2 5, 5)", &configuration),
+        eval_with_context("avg(sub2 5, 5)", &context),
         Ok(Value::Int(4))
     );
     assert_eq!(
-        eval_with_configuration("sub2(avg(3, 6))", &configuration),
+        eval_with_context("sub2(avg(3, 6))", &context),
         Ok(Value::Int(2))
     );
     assert_eq!(
-        eval_with_configuration("sub2 avg(3, 6)", &configuration),
+        eval_with_context("sub2 avg(3, 6)", &context),
         Ok(Value::Int(2))
     );
     assert_eq!(
-        eval_with_configuration("muladd(3, 6, -4)", &configuration),
+        eval_with_context("muladd(3, 6, -4)", &context),
         Ok(Value::Int(14))
     );
     assert_eq!(
-        eval_with_configuration("count()", &configuration),
+        eval_with_context("count()", &context),
         Err(EvalexprError::WrongOperatorArgumentAmount {
             actual: 0,
             expected: 1
         })
     );
     assert_eq!(
-        eval_with_configuration("count(3, 5.5, 2)", &configuration),
+        eval_with_context("count(3, 5.5, 2)", &context),
         Ok(Value::Int(3))
     );
     assert_eq!(
-        eval_with_configuration("count 5", &configuration),
+        eval_with_context("count 5", &context),
         Ok(Value::Int(1))
     );
 
     assert_eq!(
-        eval_with_configuration("min(4.0, 3)", &configuration),
+        eval_with_context("min(4.0, 3)", &context),
         Ok(Value::Int(3))
     );
     assert_eq!(
-        eval_with_configuration("max(4.0, 3)", &configuration),
+        eval_with_context("max(4.0, 3)", &context),
         Ok(Value::Float(4.0))
     );
 }
@@ -341,29 +341,29 @@ fn test_no_panic() {
 
 #[test]
 fn test_shortcut_functions() {
-    let mut configuration = HashMapContext::new();
-    configuration.set_value("string", Value::from("a string")).unwrap();
+    let mut context = HashMapContext::new();
+    context.set_value("string", Value::from("a string")).unwrap();
 
     // assert_eq!(eval_string("???"));
     assert_eq!(
-        eval_string_with_configuration("string", &configuration),
+        eval_string_with_context("string", &context),
         Ok("a string".to_string())
     );
     assert_eq!(eval_float("3.3"), Ok(3.3));
     assert_eq!(
-        eval_float_with_configuration("3.3", &configuration),
+        eval_float_with_context("3.3", &context),
         Ok(3.3)
     );
     assert_eq!(eval_int("3"), Ok(3));
-    assert_eq!(eval_int_with_configuration("3", &configuration), Ok(3));
+    assert_eq!(eval_int_with_context("3", &context), Ok(3));
     assert_eq!(eval_boolean("true"), Ok(true));
     assert_eq!(
-        eval_boolean_with_configuration("true", &configuration),
+        eval_boolean_with_context("true", &context),
         Ok(true)
     );
     assert_eq!(eval_tuple("3,3"), Ok(vec![Value::Int(3), Value::Int(3)]));
     assert_eq!(
-        eval_tuple_with_configuration("3,3", &configuration),
+        eval_tuple_with_context("3,3", &context),
         Ok(vec![Value::Int(3), Value::Int(3)])
     );
 
@@ -371,21 +371,21 @@ fn test_shortcut_functions() {
     assert_eq!(
         build_operator_tree("string")
             .unwrap()
-            .eval_string_with_configuration(&configuration),
+            .eval_string_with_context(&context),
         Ok("a string".to_string())
     );
     assert_eq!(build_operator_tree("3.3").unwrap().eval_float(), Ok(3.3));
     assert_eq!(
         build_operator_tree("3.3")
             .unwrap()
-            .eval_float_with_configuration(&configuration),
+            .eval_float_with_context(&context),
         Ok(3.3)
     );
     assert_eq!(build_operator_tree("3").unwrap().eval_int(), Ok(3));
     assert_eq!(
         build_operator_tree("3")
             .unwrap()
-            .eval_int_with_configuration(&configuration),
+            .eval_int_with_context(&context),
         Ok(3)
     );
     assert_eq!(
@@ -395,7 +395,7 @@ fn test_shortcut_functions() {
     assert_eq!(
         build_operator_tree("true")
             .unwrap()
-            .eval_boolean_with_configuration(&configuration),
+            .eval_boolean_with_context(&context),
         Ok(true)
     );
     assert_eq!(
@@ -405,7 +405,7 @@ fn test_shortcut_functions() {
     assert_eq!(
         build_operator_tree("3,3")
             .unwrap()
-            .eval_tuple_with_configuration(&configuration),
+            .eval_tuple_with_context(&context),
         Ok(vec![Value::Int(3), Value::Int(3)])
     );
 }